Move Carousel and FavoritesDisplay into containers directory and update imports in index file
